Move to a version 0_1 folder within each board
