+++ keyboards/hillside/48/0_1/0_1.c
@@ -1,4 +1,4 @@
 // Copyright 2022 Michael McCoyd (@mmccoyd)
 // SPDX-License-Identifier: GPL-2.0-or-later
 
-#include "48.h"
+#include "0_1.h"
