+++ keyboards/hillside/48/0_1/rules.mk
@@ -1,0 +1,8 @@
+SPLIT_KEYBOARD = yes        # Use shared split_common code
+LTO_ENABLE = yes            # Use link time optimization for smaller firmware
+
+# If you add a haptic board,
+# enable it and set its driver here or in your keymap folder
+# The Pimoroni board's driver is DRV2605L 
+# HAPTIC_ENABLE = yes        # Enable haptic driver
+# HAPTIC_DRIVER = DRV2605L
