+++ keyboards/hillside/48/readme.md
@@ -13,11 +13,11 @@
 
 Make example for this keyboard (after setting up your build environment):
 
-    make hillside/48:default
+    make hillside/48/0_1:default
 
 Flashing example for this keyboard:
 
-    make hillside/48:default:flash
+    make hillside/48/0_1:default:flash
 
 ## Bootloader
