+++ keyboards/hillside/52/rules.mk
@@ -1,8 +1,1 @@
-SPLIT_KEYBOARD = yes        # Use shared split_common code
-LTO_ENABLE = yes            # Use link time optimization for smaller firmware
-
-# If you add a haptic board,
-# enable it and set its driver here or in your keymap folder
-# The Pimoroni board's driver is DRV2605L 
-# HAPTIC_ENABLE = yes        # Enable haptic driver
-# HAPTIC_DRIVER = DRV2605L
+DEFAULT_FOLDER = hillside/52/0_1
